Update ListCommand to display help
In order to improve the ease of use of the food guide,
users should be able to easily get the usage syntax for
any command without having to open up the help window or user guide.

A new prefix "-h" (for help) is added to facilitate this.
Appending "-h" to any command should override existing behavior
and instead display a description of the command as well as
its usage.

Similar changes will be made to all other commands,
with the exception of "help", for that is redundant.